# api/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: arca-lb Controller REST API
  description: REST API for managing VIPs and backends in arca-lb.
  version: 0.1.0
servers:
  - url: http://localhost:8080
...
paths:
...
  /api/v1/vips:
    get:
      tags: [VIPs]
      summary: List VIPs
      operationId: listVIPs
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VIPListResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      tags: [VIPs]
      summary: Create VIP
      operationId: createVIP
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VIPCreateRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VIP"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/vips/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    get:
      tags: [VIPs]
      summary: Get VIP
      operationId: getVIP
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VIP"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      tags: [VIPs]
      summary: Update VIP
      operationId: updateVIP
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VIPUpdateRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VIP"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      tags: [VIPs]
      summary: Delete VIP
      operationId: deleteVIP
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/backends:
    get:
      tags: [Backends]
      summary: List backends for a VIP
      operationId: listBackends
      parameters:
        - name: vip_id
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BackendListResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      tags: [Backends]
      summary: Create backend
      operationId: createBackend
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BackendCreateRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Backend"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: VIP not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    Protocol:
      type: string
      enum: [TCP, UDP]
    LBMethod:
      type: string
      enum: [maglev]
    EncapType:
      type: string
      enum: [GRE4, GRE6, L3DSR, NAT4, NAT6]
    IPAddress:
      type: string
      description: IPv4 or IPv6 address.
    VIP:
      type: object
      properties:
        id:
          type: string
        vip:
          $ref: "#/components/schemas/IPAddress"
        port:
          type: integer
          minimum: 1
          maximum: 65535
        protocol:
          $ref: "#/components/schemas/Protocol"
        lb_method:
          $ref: "#/components/schemas/LBMethod"
        encap_type:
          $ref: "#/components/schemas/EncapType"
        dscp:
          type: integer
          minimum: 1
          maximum: 63
          nullable: true
        health_check:
          $ref: "#/components/schemas/HealthCheck"
        backends:
          type: array
          items:
            $ref: "#/components/schemas/Backend"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required: [id, vip, port, protocol, lb_method, created_at, updated_at]
    HealthCheck:
      type: object
      description: Health check settings. Timing is currently stored as whole seconds; millisecond-granularity timing is planned for a future API/model revision.
      properties:
        id:
          type: string
        vip_id:
          type: string
        type:
          type: string
          enum: [http, https, tcp, ping, tls-hello]
        interval_sec:
          type: integer
          minimum: 1
          maximum: 2147483647
          description: Probe interval in whole seconds.
        timeout_sec:
          type: integer
          minimum: 1
          maximum: 2147483647
          description: Probe timeout in whole seconds.
        rise_count:
          type: integer
          minimum: 1
          maximum: 2147483647
        fall_count:
          type: integer
          minimum: 1
          maximum: 2147483647
        config:
          type: object
          additionalProperties: true
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required: [id, vip_id, type, interval_sec, timeout_sec, rise_count, fall_count, created_at, updated_at]
    HealthCheckCreate:
      type: object
      description: Health check settings for VIP creation. Use interval_seconds and timeout_seconds for CRD-compatible whole-second timing. Legacy duration strings are still accepted through interval and timeout, but each timing value must be specified in exactly one representation.
      properties:
        type:
          type: string
          enum: [http, https, tcp, ping, tls-hello]
        interval_seconds:
          type: integer
          minimum: 1
          maximum: 2147483647
          description: Preferred interval in whole seconds.
        timeout_seconds:
          type: integer
          minimum: 1
          maximum: 2147483647
          description: Preferred timeout in whole seconds. Must be less than the interval.
        interval:
          type: string
          deprecated: true
          description: Legacy Go duration string that must resolve to positive whole seconds, at most 2147483647 seconds (e.g. "10s", "1m"). Do not set together with interval_seconds.
        timeout:
          type: string
          deprecated: true
          description: Legacy Go duration string that must resolve to positive whole seconds, at most 2147483647 seconds (e.g. "3s", "1m"). Do not set together with timeout_seconds.
        rise_count:
          type: integer
          minimum: 1
          maximum: 2147483647
          description: Consecutive successes to mark backend Up (default 3).
        fall_count:
          type: integer
          minimum: 1
          maximum: 2147483647
          description: Consecutive failures to mark backend Down (default 2).
        config:
          type: object
          description: Prober-specific configuration; required for http/https/tcp/tls-hello.
          additionalProperties: true
      oneOf:
        - required: [type, interval_seconds, timeout_seconds]
        - required: [type, interval, timeout]
    VIPCreateRequest:
      type: object
      properties:
        vip:
          $ref: "#/components/schemas/IPAddress"
        port:
          type: integer
          minimum: 1
          maximum: 65535
        protocol:
          $ref: "#/components/schemas/Protocol"
        lb_method:
          $ref: "#/components/schemas/LBMethod"
        encap_type:
          $ref: "#/components/schemas/EncapType"
        dscp:
          type: integer
          minimum: 1
          maximum: 63
          nullable: true
        health_check:
          $ref: "#/components/schemas/HealthCheckCreate"
      required: [vip, port, protocol]
    VIPUpdateRequest:
      type: object
      properties:
        vip:
          $ref: "#/components/schemas/IPAddress"
        port:
          type: integer
          minimum: 1
          maximum: 65535
        protocol:
          $ref: "#/components/schemas/Protocol"
        lb_method:
          $ref: "#/components/schemas/LBMethod"
        encap_type:
          $ref: "#/components/schemas/EncapType"
        dscp:
          type: integer
          minimum: 1
          maximum: 63
          nullable: true
      additionalProperties: false
    VIPListResponse:
      type: object
      properties:
        vips:
          type: array
          items:
            $ref: "#/components/schemas/VIP"
        count:
          type: integer
      required: [vips, count]
    Backend:
      type: object
      properties:
        id:
          type: string
        vip_id:
          type: string
        ip:
          $ref: "#/components/schemas/IPAddress"
        weight:
          type: integer
          minimum: 1
          maximum: 100
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required: [id, vip_id, ip, weight, created_at, updated_at]
    BackendCreateRequest:
      type: object
      properties:
        vip_id:
          type: string
        ip:
          $ref: "#/components/schemas/IPAddress"
        weight:
          type: integer
          minimum: 1
          maximum: 100
      required: [vip_id, ip]
...
    BackendListResponse:
      type: object
      properties:
        backends:
          type: array
          items:
            $ref: "#/components/schemas/Backend"
        count:
          type: integer
      required: [backends, count]